+++ Project/classes.h
@@ -23,16 +23,16 @@ class elevator;
 
 class visitor 
 {
-  int floor;
-  int office_num;
-  int priority; 
+  short floor;
+  short office_num;
+  unsigned int priority; 
   bool is_satisfied;
 public:
   visitor(int fl, int off);
   void set_priority(int);
-  int get_priority();
-  int get_office_num();
-  int get_floor();
+  unsigned int get_priority();
+  short get_office_num();
+  short get_floor();
   void set_satisfaction(bool);
   bool get_satisfaction(); 
 };
@@ -41,14 +41,14 @@ public:
 
 class waiting_room
 {
-  int curr;
+  unsigned int curr;
   queue<visitor*> visitors;  
 public:
   waiting_room();    /* Waiting room does not have maximum capacity */
   ~waiting_room();
   void enter(visitor*);
   visitor* exit(); 
-  int get_curr();
+  unsigned int get_curr();
   queue<visitor*> get_vst();
 };
 
@@ -56,8 +56,8 @@ public:
 
 class ground_level
 {
-  int cap;
-  int curr;
+  unsigned int cap;
+  unsigned int curr;
   building* bld;  
   waiting_room* wr;
 public:
@@ -66,8 +66,8 @@ public:
   bool enter(visitor*); 
   void exit(visitor*); 
   void wait(visitor*); //metaferei ton visitor sto wr 
-  int get_cap();
-  int get_curr();
+  unsigned int get_cap();
+  unsigned int get_curr();
   waiting_room* get_wr();
 };
 
@@ -75,16 +75,16 @@ public:
 
 class office
 {
-  int number;
-  int cap;
-  int total; // total visitors, used to prioritize ppl (bank-style)       
+  short number;
+  unsigned int cap;
+  unsigned int total; // total visitors, used to prioritize ppl (bank-style)       
   queue<visitor*> visitors;
 public:
   office(int No, int num);
   ~office();
   bool enter(visitor *);
   visitor *exit();
-  int get_cap();
+  unsigned int get_cap();
   bool is_empty();  //[Harry] used in elev::stop_down
 };
 
@@ -92,9 +92,9 @@ public:
 
 class floor
 {
-  int number; // [1,4]
-  int cap; 
-  int curr;
+  short number; // [1,4]
+  unsigned int cap; 
+  unsigned int curr;
   waiting_room* wr;
   office** off;
 public:
@@ -102,8 +102,8 @@ public:
   ~floor();
   bool enter(visitor *);    
   visitor *exit();
-  int get_cap();
-  int get_curr();
+  unsigned int get_cap();
+  unsigned int get_curr();
   waiting_room* get_wr();
   office *get_office(int office_n); //[Harry] added this one used on elevator::stop_down
 };
@@ -112,13 +112,13 @@ public:
 
 class elevator
 {
-  int total; // [Harry] used to prioritize ppl as usual
-  int cap;
-  int curr_fl;
+  unsigned int total; // [Harry] used to prioritize ppl as usual
+  unsigned int cap;
+  short curr_fl;
   floor** fl;
   ground_level* grl;
-  int curr;
-  int crcl_rem;   // circles remaining // ousiastika termatizei th diadikasia
+  unsigned int curr;
+  unsigned int crcl_rem;   // circles remaining // ousiastika termatizei th diadikasia
   queue<visitor*> visitors;
   bool enter(visitor*);     
   void exit(visitor*);   
@@ -129,16 +129,16 @@ public:
   elevator(int Nl, int lc, floor **,ground_level*);    
   ~elevator();
   void operate();
-  int get_cap();
-  int get_curr();
+  unsigned int get_cap();
+  unsigned int get_curr();
 };
 
 /* ===========================================================||  B U I L D I N G  ||=========================================================== */
 
 class building
 {
-  int cap;  /* capacity */
-  int curr; /*  current ppl inside */
+  unsigned int cap;  /* capacity */
+  unsigned int curr; /*  current ppl inside */
   ground_level* gr_lvl;
   floor** fl;  // Floor pointer (create floors dynamically during construction)
   elevator* el;
